filmatic_backend-master/requirements.txt
fastapi
joblib
motor
pandas
nltk
pydantic
pymongo
python-dotenv
scikit_learn
uvicorn